kafka/API/requirements.txt
newsapi-python>=0.2.7
python-dotenv>=1.0.0
